Conflicts:
	app/controllers/topics_controller.rb
	app/policies/application_policy.rb
	app/views/topics/index.html.erb
	app/views/welcome/index.html.erb

+++ app/controllers/topics_controller.rb
@@ -1,13 +1,11 @@
 class TopicsController < ApplicationController
   def index
     @topics = Topic.all
-    authorize @topics
   end
 
   def show
     @topic = Topic.find(params[:id])
     @bookmarks = @topic.bookmarks
-    authorize @topic
   end
 
   def new
@@ -17,18 +15,30 @@ class TopicsController < ApplicationController
 
   def edit
     @topic = Topic.find(params[:id])
+    authorize @topic
   end
 
   def create
-
-    @topic = Topic.new(topic_params)
+    @topic = current_user.topics.build(topic_params)
     authorize @topic
 
     if @topic.save
-      redirect_to @topic, notice: "Topic was Saved"
+      redirect_to @topic, notice: "Topic was Added"
     else
       flash[:error] = "There was an error creating Topic. Please try again"
       render :new
+    end
+  end
+
+  def update
+    @topic = Topic.find(params[:id])
+    authorize @topic
+
+    if @topic.update_attributes(topic_params)
+      redirect_to @topic
+    else
+      flash[:error] = "Error saving, try again"
+      render :edit
     end
   end
 
@@ -36,10 +46,10 @@ class TopicsController < ApplicationController
   #create and destroy from the topics#index view
   def destroy
     @topic = Topic.find(params[:id])
-    name = @topic.name
-    #authorize @topics
+    #name = @topic.title
+    authorize @topic
     if @topic.destroy
-      flash[:notice] ="\#{name}\" was deleted."
+      flash[:notice] ="\#{title}\" topic was deleted."
       redirect_to topics_show
     else
       flash[:error] = "There was an error deleting"
@@ -49,6 +59,6 @@ class TopicsController < ApplicationController
 
   private
   def topic_params
-    params.require(:topic).permit(:name)
+    params.require(:topic).permit(:title)
   end
 end
